Create styling for on addcar through car.css
