+++ src/components/Addcar/Car.css
@@ -1,0 +1,64 @@
+body,
+html {
+  /* background-color: green; */
+  /* background: linear-gradient(to center, #ee0505 50%, #1ee91e 50%); */
+  /* width: 400px; */
+  display: flex;
+  height: 100%;
+  width: 100%;
+  align-items: center;
+  justify-content: center;
+  /* position: absolute; */
+}
+
+.innercontainer {
+  background-color: aqua;
+  width: 40vw;
+  height: 90vh;
+  /* margin-top: 50px; */
+  padding-top: 30px;
+  border-radius: 8px;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  max-width: 400px;
+  margin: auto;
+}
+
+label {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 1rem;
+}
+
+label > span {
+  font-size: 1.2rem;
+  font-weight: bold;
+  margin-bottom: 0.5rem;
+}
+
+input[type="text"],
+textarea,
+input[type="file"] {
+  padding: 0.5rem;
+  border-radius: 4px;
+  border: 1px solid #ccc;
+  font-size: 1rem;
+}
+
+button[type="submit"] {
+  background-color: hsl(0, 85%, 47%);
+  color: #fff;
+  padding: 0.5rem 1rem;
+  border: none;
+  border-radius: 4px;
+  font-size: 1rem;
+  cursor: pointer;
+  margin-top: 1rem;
+}
+
+button[type="submit"]:hover {
+  background-color: #0030cc;
+}
